Add shared adapter generators
